requirements.txt
PyQt5==5.15.9
mysql-connector-python==8.0.33
pytest==7.4.0
python-dotenv==1.0.0
bcrypt==4.0.1
pydantic==1.10.9